refactor: 누락된 export 반영, props로 받는 type 분리
